Removed user32.dll dependency for mouse movement
This will change how mouse movement is computed on windows. This was here from before to solve a bug with double clicking on an empty textbox. That change was pretty large and the description was lacking. I'm assuming this is because I couldn't find this specific function at that time. This change also has survived some RichTextBox changes so I'm hoping it's related specifically to that. I'm betting I had a method to do it on a RichTextBox and found find the equivalent on stackoverflow for TextBox first as Windows messages. Either way, if mouse movement degrades, this is the change for you.